+++ django_couchdb_utils/auth/tests.py
