+++ django_couchdb_utils/cache/_design/views/cache_by_key/map.js
