+++ django_couchdb_utils/email/backends.py
@@ -9,28 +9,6 @@ DEFAULT_WRAPPED_BACKEND='django.core.mail.backends.smtp.EmailBackend'
 COPY_PROPERTIES=('subject', 'body', 'from_email', 'to', 'cc', 'bcc', 'headers')
 
 
-class EmailMessage(Document):
-    """
-    A couchdbkit Document to store emails, uses the same properties as
-    django.core.mail.EmailMessage
-    """
-
-    subject     = StringProperty()
-    body        = StringProperty()
-    from_email  = StringProperty(default=settings.DEFAULT_FROM_EMAIL)
-    to          = StringListProperty()
-    cc          = StringListProperty()
-    bcc         = StringListProperty()
-#   attachments = blobs/attachments ?
-    headers     = DictProperty()
-
-    @classmethod
-    def all_messages(cls):
-        r = cls.view('django_couchdb_utils/emails', include_docs=True)
-        return list(r)
-
-    def __repr__(self):
-        return 'EmailMessage (%s)' % self._id
 
 
 class CouchDBEmailBackend(BaseEmailBackend):
