+++ django_couchdb_utils/sessions/_design/views/sessions_by_key/map.js
