+++ django_couchdb_utils/sessions/models.py
@@ -1,0 +1,10 @@
+class Session(Document):
+    session_key  = StringProperty()
+    session_data = StringProperty()
+    expire_date  = DateTimeProperty()
+
+    @classmethod
+    def get_session(cls, session_key):
+        r = cls.view('django_couchdb_utils/sessions_by_key', key=session_key, include_docs=True)
+        return r.first() if r else None
+
