+++ django_couchdb_utils/sessions/tests.py
