+++ django_couchdb_utils/test/utils.py
